requirements.txt
aiohttp==3.7.4.post0
pyserial==3.5
pyserial-asyncio==0.5
aiohttp-xmlrpc==1.3.0
progressbar2==3.53.1
requests==2.26.0
networkx
progressbar
numpy
